WFCORE-6808 Add capability-resolving variant of child service target-based ServiceInstaller.Builder
