+++ server/src/main/java/org/jboss/as/server/deployment/DeploymentUnitPhaseService.java
@@ -13,26 +13,20 @@ import java.util.ListIterator;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
-import org.jboss.as.controller.RequirementServiceBuilder;
 import org.jboss.as.controller.RequirementServiceTarget;
-import org.jboss.as.controller.capability.CapabilityServiceSupport;
 import org.jboss.as.server.deployment.module.ModuleSpecification;
 import org.jboss.as.server.logging.ServerLogger;
 import org.jboss.modules.ModuleIdentifier;
 import org.jboss.msc.service.LifecycleEvent;
 import org.jboss.msc.service.LifecycleListener;
-import org.jboss.msc.service.DelegatingServiceBuilder;
 import org.jboss.msc.service.DelegatingServiceRegistry;
-import org.jboss.msc.service.DelegatingServiceTarget;
 import org.jboss.msc.service.Service;
 import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceContainer;
 import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.ServiceController.Mode;
 import org.jboss.msc.service.ServiceName;
-import org.jboss.msc.service.ServiceTarget;
 import org.jboss.msc.service.StartContext;
 import org.jboss.msc.service.StartException;
 import org.jboss.msc.service.StopContext;
@@ -109,7 +103,7 @@ final class DeploymentUnitPhaseService<T> implements Service<T> {
         final List<RegisteredDeploymentUnitProcessor> list = chains.getChain(phase);
         final ListIterator<RegisteredDeploymentUnitProcessor> iterator = list.listIterator();
         final ServiceContainer container = context.getController().getServiceContainer();
-        final RequirementServiceTarget serviceTarget = new DeploymentUnitServiceTarget(context.getChildTarget().subTarget(), deploymentUnit.getAttachment(Attachments.CAPABILITY_SERVICE_SUPPORT));
+        final RequirementServiceTarget serviceTarget = RequirementServiceTarget.forTarget(context.getChildTarget().subTarget(), deploymentUnit.getAttachment(Attachments.CAPABILITY_SERVICE_SUPPORT));
         final DeploymentUnit parent = deploymentUnit.getParent();
 
         final List<Consumer<ServiceBuilder<?>>> dependencies = new LinkedList<>();
@@ -266,68 +260,5 @@ final class DeploymentUnitPhaseService<T> implements Service<T> {
             }
         }
         return !shouldNotRun.contains(deployer.getSubsystemName());
-    }
-
-    static class DeploymentUnitServiceTarget extends DelegatingServiceTarget implements RequirementServiceTarget {
-        private final CapabilityServiceSupport support;
-
-        DeploymentUnitServiceTarget(ServiceTarget target, CapabilityServiceSupport support) {
-            super(target);
-            this.support = support;
-        }
-
-        @Override
-        public RequirementServiceTarget addListener(LifecycleListener listener) {
-            super.addListener(listener);
-            return this;
-        }
-
-        @Override
-        public RequirementServiceTarget removeListener(LifecycleListener listener) {
-            super.removeListener(listener);
-            return this;
-        }
-
-        @Override
-        public RequirementServiceTarget subTarget() {
-            return new DeploymentUnitServiceTarget(super.subTarget(), this.support);
-        }
-
-        @Override
-        public RequirementServiceBuilder<?> addService() {
-            return new DeploymentUnitServiceBuilder<>(super.addService(), this.support);
-        }
-    }
-
-    static class DeploymentUnitServiceBuilder<T> extends DelegatingServiceBuilder<T> implements RequirementServiceBuilder<T> {
-        private final CapabilityServiceSupport support;
-
-        DeploymentUnitServiceBuilder(ServiceBuilder<T> builder, CapabilityServiceSupport support) {
-            super(builder);
-            this.support = support;
-        }
-
-        @Override
-        public RequirementServiceBuilder<T> setInitialMode(Mode mode) {
-            super.setInitialMode(mode);
-            return this;
-        }
-
-        @Override
-        public RequirementServiceBuilder<T> setInstance(org.jboss.msc.Service service) {
-            super.setInstance(service);
-            return this;
-        }
-
-        @Override
-        public RequirementServiceBuilder<T> addListener(LifecycleListener listener) {
-            super.addListener(listener);
-            return this;
-        }
-
-        @Override
-        public <V> Supplier<V> requiresCapability(String capabilityName, Class<V> dependencyType, String... referenceNames) {
-            return super.requires(this.support.getCapabilityServiceName(capabilityName, referenceNames));
-        }
     }
 }
